pyproject.toml
[tool.poetry]
name = "git-888"
version = "0.1.0"
description = ""
authors = ["Pic <nik20010106@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.1.1"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
